Renamed `astro.files.type` to `astro.files.types`
Renamed the file for astro.files.type to astro.files.types
