+++ src/astro/files/types/ndjson.py
@@ -6,7 +6,7 @@ import pandas as pd
 
 from astro.constants import DEFAULT_CHUNK_SIZE
 from astro.constants import FileType as FileTypeConstants
-from astro.files.type.base import FileType
+from astro.files.types.base import FileType
 
 
 class NDJSONFileType(FileType):
