+++ src/astro/files/types/parquet.py
@@ -3,7 +3,7 @@ import io
 import pandas as pd
 
 from astro.constants import FileType as FileTypeConstants
-from astro.files.type.base import FileType
+from astro.files.types.base import FileType
 
 
 class ParquetFileType(FileType):
