+++ tests/files/type/test_type_base.py
@@ -3,7 +3,7 @@ import pathlib
 import pytest
 
 from astro.constants import FileType
-from astro.files.type import get_filetype
+from astro.files.types import get_filetype
 
 sample_file = pathlib.Path(pathlib.Path(__file__).parent.parent, "data/sample.csv")
 sample_filepaths_per_filetype = [
